feat(cli): add `list` logic

+++ src/db.rs
@@ -3,7 +3,9 @@ use sqlx::sqlite::{SqlitePool, SqlitePoolOptions};
 pub async fn connect(conn: &str, max_conn: Option<u32>) -> Result<SqlitePool, sqlx::Error> {
     let max_conn = max_conn.unwrap_or(5);
 
-    std::fs::File::create(conn)?;
+    if !std::path::Path::new(conn).exists() {
+        std::fs::File::create(conn)?;
+    }
 
     let c = SqlitePoolOptions::new()
         .max_connections(max_conn)
